Finish logic for Animation module, mostly
Done with most of the logic for the Animation module in src/Animation.jl (responsible for creating a Scene object and updating it with each discrete time step of the walk, later sent to the server by Routes). Potentially subject to updates/added functionality in the future. Docs and tests will be added in a future commit.